nwash_backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.0
django-cors-headers==4.4.0
gunicorn==22.0.0
psycopg2>=2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
Pillow>=10.3.0
drf-yasg
